Put option handling in a class so it'll be easier to use for Reel
