+++ lib/pacto/server.rb
@@ -1,4 +1,5 @@
 # -*- encoding : utf-8 -*-
 require 'goliath'
 require 'pacto'
+require 'pacto/server/settings'
 require 'pacto/server/api'
